edited readme, because of this is a fork copied my edited file into the original one removed comments from mcuchronologicalplaylist.txt
